refactor: isolate install logic
